fix possible integer overflows in timing functions

+++ src/i_video.c
@@ -895,7 +895,7 @@ void I_FinishUpdate (void)
     // [AM] Figure out how far into the current tic we're in as a fixed_t.
     if (crispy->uncapped)
     {
-	fractionaltic = I_GetTimeMS() * TICRATE % 1000 * FRACUNIT / 1000;
+	fractionaltic = (int64_t)I_GetTimeMS() * TICRATE % 1000 * FRACUNIT / 1000;
     }
 
     // Restore background and undo the disk indicator, if it was drawn.
